backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
pikepdf
Pillow
python-docx
openpyxl
qrcode[pil]
python-dotenv
